Settle on cmd2
