--- llmsat/components/__init__.py
@@ -1,3 +1,0 @@
-from llmsat.components.eps import EPS
-from llmsat.components.obc import OBC
-from llmsat.components.payload import PayloadManager
